move tests to pkg/tests
