+++ pkg/tests/alertmanager/grafana.go
@@ -1,4 +1,4 @@
-package testutil
+package alertmanager
 
 import (
 	_ "embed"
