+++ pkg/tests/alertmanager/postgres.go
@@ -1,4 +1,4 @@
-package testutil
+package alertmanager
 
 import (
 	"os"
